count number of seconds a button is pressed
Count number of seconds a button is pressed. This gives the option to react differently on a long button press. User visible change: The effect of the button press is now only coming into effect when the button has been released - before the effect of the button press was happening as soon as the button was pressed.